Clustering4D: add casts

+++ src/modules/Clustering4D/Clustering4D.cpp
@@ -175,8 +175,8 @@ StatusCode Clustering4D::run(const std::shared_ptr<Clipboard>& clipboard) {
 
         // Fill cluster histograms
         clusterSize->Fill(static_cast<double>(cluster->size()));
-        clusterWidthRow->Fill(cluster->rowWidth());
-        clusterWidthColumn->Fill(cluster->columnWidth());
+        clusterWidthRow->Fill(static_cast<double>(cluster->rowWidth()));
+        clusterWidthColumn->Fill(static_cast<double>(cluster->columnWidth()));
         clusterCharge->Fill(cluster->charge());
         if(cluster->size() == 1) {
             clusterCharge_1px->Fill(cluster->charge());
